Restore the correct version of CLHEP to work with the Mu2e Offline master branch, in particular version v6_0_3 and v6_0_4.  This is a custom version of clhep: v2_2_0_8a.
The head of BTrk had been configured to work with the g4v10-validation-branch
of Mu2e Offline. That will be moved to a branch of BTrk.

The history of recent tags on the BTrk master branch is:

v1_01_03 - works with Offline master v6_0_3 and v6_0_4 and nearby
v1_01_04 - works with Offline g4v10-validation-branch
v1_01_05 - Bug fix to resolve physical/logical path confusion
v1_01_06 - works with Offline master v6_0_3 and v6_0_4 and nearby